requirements.txt
python-dotenv
langchain
openai
instructor
pydantic
fastapi
setuptools
uvicorn
#openai-whisper
